Revert "“iteration-3”"
This reverts commit 42323ea13954e677a802262be33cc0b4a666825f.

+++ ydb/tests/olap/ttl_tiering/ttl_delete_s3.py
@@ -15,6 +15,12 @@ class TllDeleteBase(TllTieringTestBase):
     days_to_freeze = 3000
 
     @classmethod
+    def portions_actualized_in_sys(self, table):
+        portions = table.get_portion_stat_by_tier()
+        logger.info(f"portions: {portions}, blobs: {table.get_blob_stat_by_tier()}")
+        return "__DEFAULT" in portions and self.row_count > portions["__DEFAULT"]["Rows"]
+
+    @classmethod
     def get_row_count_by_date(self, table_path: str, past_days: int) -> int:
         return self.ydb_client.query(f"SELECT count(*) as Rows from `{table_path}` WHERE ts < CurrentUtcTimestamp() - DateTime::IntervalFromDays({past_days})")[0].rows[0]["Rows"]
 
@@ -267,7 +273,8 @@ class TestDeleteS3Ttl(TllDeleteBase):
         logger.info(f"Rows older than {self.days_to_cool} days: {self.get_row_count_by_date(table_path, self.days_to_cool)}")
         logger.info(f"Rows older than {self.days_to_freeze} days: {self.get_row_count_by_date(table_path, self.days_to_freeze)}")
 
-        assert ColumnTableHelper.portions_actualized_in_sys(self.table)
+        if not self.wait_for(lambda: self.portions_actualized_in_sys(self.table), 120):
+            raise Exception(".sys reports incorrect data portions")
 
         stmt = f"""
             DELETE FROM `{table_path}`
@@ -275,7 +282,9 @@ class TestDeleteS3Ttl(TllDeleteBase):
         logger.info(stmt)
         self.ydb_client.query(stmt)
 
-        assert self.wait_for(lambda: self.data_deleted_from_buckets('cold_delete', 'frozen_delete'), 200), "not all data deleted"
+        if not self.wait_for(lambda: self.data_deleted_from_buckets('cold_delete', 'frozen_delete'), 300):
+            # raise Exception("not all data deleted") TODO FIXME after https://github.com/ydb-platform/ydb/issues/13594
+            pass
 
 
 class TestDeleteTtl(TllDeleteBase):
@@ -373,7 +382,8 @@ class TestDeleteTtl(TllDeleteBase):
         logger.info(f"Rows older than {self.days_to_cool} days: {self.get_row_count_by_date(table_path, self.days_to_cool)}")
         logger.info(f"Rows older than {self.days_to_freeze} days: {self.get_row_count_by_date(table_path, self.days_to_freeze)}")
 
-        assert ColumnTableHelper.portions_actualized_in_sys(self.table)
+        if not self.wait_for(lambda: self.portions_actualized_in_sys(self.table), 200):
+            raise Exception(".sys reports incorrect data portions")
 
         t0 = time.time()
         stmt = f"""
@@ -398,4 +408,5 @@ class TestDeleteTtl(TllDeleteBase):
             # So we wait until some data appears in any bucket
             return cold_bucket_stat[0] != 0 or frozen_bucket_stat[0] != 0
 
-        assert self.wait_for(lambda: data_distributes_across_tiers(), 200), "Data eviction has not been started"
+        if not self.wait_for(lambda: data_distributes_across_tiers(), 600):
+            raise Exception("Data eviction has not been started")
